backend/requirements.txt
# Vision Agents SDK with all plugins
vision-agents[getstream,gemini,ultralytics,elevenlabs,deepgram]

# Web server
uvicorn[standard]>=0.38.0

# Environment variables
python-dotenv>=1.1.1

# Math for keypoint calculations
numpy>=1.26.4

# Stream server-side SDK for token generation
stream-chat>=4.10.0

# FastAPI
fastapi>=0.115.0